Add authorization functionality

+++ nodejs-api/backend/routes/posts.js
@@ -4,6 +4,7 @@ const checkAuth = require("../middleware/check-auth");
 
 const Post = require('../models/post');
 const { utimes } = require("fs");
+const { use } = require("./user");
 
 const router = express.Router();
 
@@ -37,7 +38,8 @@ router.post('', checkAuth, multer({ storage: storage }).single("image"), (req, r
     const post = new Post({
         title: req.body.title,
         content: req.body.content,
-        imagePath: url + '/images/' + req.file.filename
+        imagePath: url + '/images/' + req.file.filename,
+        creator: req.userData.userId
     });
     post.save().then(result => {
         res.status(201).json({
@@ -96,11 +98,14 @@ router.get('', (req, res, next) => {
 });
 
 router.delete('/:id', checkAuth, (req, res, next) => {
-    Post.deleteOne({ _id: req.params.id }).then(result => {
-        console.log(result);
-        res.status(200).json({
-            message: "Post deleted successfully"
-        });
+    Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+        // console.log(result);
+        if (result.n > 0) {
+            res.status(200).json({ message: "Post deleted successfully" });
+        }
+        else {
+            res.status(401).json({ message: "Not authorized!" });
+        }
     });
 });
 
@@ -114,11 +119,17 @@ router.put("/:id", checkAuth, multer({ storage: storage }).single("image"), (req
         _id: req.params.id,
         title: req.body.title,
         content: req.body.content,
-        imagePath: imagePath
+        imagePath: imagePath,
+        creator: req.userData.userId
     });
-    Post.updateOne({ _id: req.params.id }, post).then(result => {
-        console.log(result);
-        res.status(200).json({ message: "Update successful!" });
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+        // console.log(result);
+        if (result.nModified > 0) {
+            res.status(200).json({ message: "Update successful!" });
+        }
+        else {
+            res.status(401).json({ message: "Not authorized!" });
+        }
     });
 });
